chore: clean up checks endpoints
