+++ apps/webservice/src/app/[workspaceSlug]/(app)/(deploy)/(raw)/systems/[systemSlug]/(raw)/deployments/[deploymentSlug]/(raw)/releases/[releaseId]/checks/_components/flow-diagram/checks/VersionSelector.tsx
@@ -5,13 +5,14 @@ export const VersionSelectorCheck: React.FC<{
   versionId: string;
   environmentId: string;
 }> = ({ versionId, environmentId }) => {
-  const { data, isLoading } =
-    api.deployment.version.checks.versionSelector.useQuery({
-      versionId,
-      environmentId,
-    });
+  const { data, isLoading } = api.policy.evaluate.useQuery({
+    environmentId,
+    versionId,
+  });
 
-  const isPassingVersionSelector = data ?? false;
+  const isPassingVersionSelector = Object.values(
+    data?.rules.versionSelector ?? {},
+  ).every((isPassing) => isPassing);
 
   if (isLoading)
     return (
