--- packages/api/src/router/deployment-version-checks/deny-window.ts
@@ -1,24 +1,0 @@
-import { z } from "zod";
-
-import { Permission } from "@ctrlplane/validators/auth";
-
-import { createTRPCRouter, protectedProcedure } from "../../trpc";
-
-export const denyWindowRouter = createTRPCRouter({
-  status: protectedProcedure
-    .input(
-      z.object({
-        workspaceId: z.string().uuid(),
-        versionId: z.string().uuid(),
-        environmentId: z.string().uuid(),
-      }),
-    )
-    .meta({
-      authorizationCheck: ({ canUser, input }) =>
-        canUser.perform(Permission.DeploymentVersionGet).on({
-          type: "deploymentVersion",
-          id: input.versionId,
-        }),
-    })
-    .query(() => false),
-});
